+++ src/post/post.module.ts
@@ -6,6 +6,7 @@ import { UserModule } from 'src/user/user.module';
 import { Post } from 'src/entities/post.entity';
 import { Tag} from 'src/entities/tag.entity'
 import { User } from 'src/entities/user.entity';
+import { TagService } from './tag.service';
 
 @Module({
   imports: [
@@ -13,6 +14,6 @@ import { User } from 'src/entities/user.entity';
     UserModule
   ],
   controllers: [ PostController ],
-  providers: [ PostService ]
+  providers: [ PostService, TagService ]
 })
 export class PostModule {}
