+++ src/post/post.service.ts
@@ -9,9 +9,7 @@ import { DeepPartial, LessThan, Repository } from 'typeorm';
 export class PostService {
   constructor(
     @InjectRepository(Post)
-    private readonly postRepository: Repository<Post>,
-    @InjectRepository(Tag)
-    private readonly tagRepository: Repository<Tag>,
+    private readonly postRepository: Repository<Post>
   ){}
   /**
    * @param definition
@@ -78,40 +76,6 @@ export class PostService {
     }
 
     return { data: posts, next: posts[posts.length - 1].createdAt }
-  }
-  /**
-   * @param  searchQuery
-   * filters to find a post
-   * @param  {Date} cursor
-   * date before post was created(provided as next)
-   * @param {number} limit
-   * post objects quantity
-   */
-  async getPostsByTag(searchQuery, cursor: Date, limit: number){
-    
-    const { tag, ...restFilters } = searchQuery;
-    
-    const tagData = await this.tagRepository
-    .createQueryBuilder('tag')
-    .where({ name: tag, ...restFilters })
-    .innerJoinAndSelect('tag.posts','posts')
-    .andWhere("posts.createdAt < :cursorDate", { cursorDate: cursor })
-    .select(["tag", "posts"])
-    .orderBy('posts.createdAt', 'DESC')
-    .limit(limit)
-    .getOne()
-
-    if (!tagData || !tagData.posts) {
-      return { data: [], next: null }
-    }
-    
-    const { posts } = tagData;
-
-    if( posts.length < limit){
-      return { data: posts, next: null }
-    }
-
-    return { data: posts, next: posts[posts.length - 1].createdAt };
   }
   /**
    * @param  searchQuery
